+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.3.3
 - King attacks are calculated in advance for mainline moves and can be accessed with `state.meta!.checks!`. They come in the form of a list of squares containing pieces that are attacking the king of each player.
+- Gating move logic is no longer part of the standard move logic. `GatingMove` is its own type which contains a child move.
 
 ### 1.3.2
 - Fixed a castling bug (queenside allowed with piece on b1).